requirements.txt
aiodns==3.2.0
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
discord.js==1.0.7
frozenlist==1.4.1
idna==3.7
javascriptpy==1.1.6
msgspec==0.18.6
multidict==6.0.5
mutagen==1.47.0
numpy==1.26.4
opencv-python==4.9.0.80
py-cord==2.5.0
pycares==4.4.0
pycparser==2.22
pycryptodomex==3.20.0
PyNaCl==1.5.0
python-dotenv==1.0.1
requests==2.31.0
typing_extensions==4.11.0
urllib3==2.2.1
websockets==12.0
yarl==1.9.4
youtube-search==2.1.2
yt-dlp==2024.4.9
